+++ src/Debug/WhoopsDebugger.php
@@ -2,9 +2,7 @@
 
 namespace Rougin\Slytherin\Debug;
 
-use Whoops\Run;
-use Whoops\Handler\HandlerInterface;
-use Whoops\Handler\PrettyPageHandler;
+use Rougin\Slytherin\Debug\Whoops\Debugger as BaseDebugger;
 
 /**
  * Debugger
@@ -17,82 +15,4 @@ use Whoops\Handler\PrettyPageHandler;
  * @package Slytherin
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
-class WhoopsDebugger implements DebuggerInterface
-{
-    /**
-     * @var string
-     */
-    protected $environment = '';
-
-    /**
-     * @var \Whoops\Run
-     */
-    protected $whoops;
-
-    /**
-     * @param \Whoops\Run $whoops
-     * @param string      $environment
-     */
-    public function __construct(Run $whoops, $environment = 'development')
-    {
-        $this->whoops = $whoops;
-        $this->environment = $environment;
-    }
-
-    /**
-     * Sets up the environment to be used.
-     * 
-     * @param  string $environment
-     * @return void
-     */
-    public function setEnvironment($environment)
-    {
-        $this->environment = $environment;
-    }
-
-    /**
-     * Gets the specified environment.
-     *
-     * @return string
-     */
-    public function getEnvironment()
-    {
-        return $this->environment;
-    }
-
-    /**
-     * Returns a listing of handlers.
-     * 
-     * @return \Whoops\Handler\HandlerInterface[]
-     */
-    public function getHandlers()
-    {
-        return $this->whoops->getHandlers();
-    }
-
-    /**
-     * Registers the instance as a debugger.
-     * 
-     * @return \Whoops\Run
-     */
-    public function display()
-    {
-        error_reporting(E_ALL);
-
-        if ($this->environment === 'development') {
-            $this->setHandler(new PrettyPageHandler);
-        }
-
-        return $this->whoops->register();
-    }
-
-    /**
-     * Sets a handler.
-     * 
-     * @param \Whoops\Handler\HandlerInterface|callable $handler
-     */
-    public function setHandler($handler)
-    {
-        $this->whoops->pushHandler($handler);
-    }
-}
+class WhoopsDebugger extends BaseDebugger {}