+++ src/IoC/AurynContainer.php
@@ -2,7 +2,7 @@
 
 namespace Rougin\Slytherin\IoC;
 
-use Auryn\Injector;
+use Rougin\Slytherin\IoC\Auryn\Container as BaseAurynContainer;
 
 /**
  * Auryn Container
@@ -15,43 +15,4 @@ use Auryn\Injector;
  * @package Slytherin
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
-class AurynContainer extends BaseContainer implements ContainerInterface
-{
-    /**
-     * @var \Auryn\Injector
-     */
-    public $injector;
-
-    /**
-     * @param \Auryn\Injector $injector
-     */
-    public function __construct(Injector $injector)
-    {
-        $this->injector = $injector;
-    }
-
-    /**
-     * Adds a new instance to the container.
-     * 
-     * @param string $id
-     * @param mixed  $concrete
-     */
-    public function add($id, $concrete = null)
-    {
-        if ($concrete && ! is_array($concrete)) {
-            $this->instances[$id] = $concrete;
-
-            return $this;
-        }
-
-        $arguments = [];
-
-        if (is_array($concrete)) {
-            $arguments = $concrete;
-        }
-
-        $this->instances[$id] = $this->injector->make($id, $arguments);
-
-        return $this;
-    }
-}
+class AurynContainer extends BaseAurynContainer {}