+++ src/IoC/Vanilla/Exception/NotFoundException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rougin\Slytherin\IoC\Exception;
+namespace Rougin\Slytherin\IoC\Vanilla\Exception;
 
 use InvalidArgumentException;
 use Interop\Container\Exception\NotFoundException as InteropNotFoundException;
@@ -13,5 +13,4 @@ use Interop\Container\Exception\NotFoundException as InteropNotFoundException;
  * @package Slytherin
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
-class NotFoundException extends InvalidArgumentException implements
-    InteropNotFoundException {}
+class NotFoundException extends InvalidArgumentException implements InteropNotFoundException {}