+++ tests/IoC/ContainerTest.php
@@ -146,7 +146,7 @@ class ContainerTest extends PHPUnit_Framework_TestCase
     public function testGetMethodWithError()
     {
         $this->setExpectedException(
-            'Rougin\Slytherin\IoC\Exception\NotFoundException'
+            'Rougin\Slytherin\IoC\Vanilla\Exception\NotFoundException'
         );
 
         $this->container->get($this->class);